feat : 당첨 번호 입력에 따른 list 생성 기능 개발
